Fix resuming from checkpoint at new stage
At the start of a new stage, optimizer, scaler, and LR-scheduler are
re-created and initialized according to the stage description.

If we resume from a checkpoint, starting off with a new stage, we thus
must not load any of that state and only load model state. Otherwise, if
the checkpoint has, for example, been created at the end of the previous
stage, the optimizer, scaler, and LR-scheduler state belong to that
stage and not the new one.

This then can (if it does not fail outright) override the initialzation
described/intended for the stage with the state from the previous stage.

Thus, when beginning a new stage, only load the model state.